Big changes! Still to clean but runnable.
I added two databases, libraries to access databases and a Hash class.

Now The Database Helper works and can connect to two different DBs ( one for password and one for salt, internet said it was better to save them in different places).
Authentication works!
If you use username:"test" and password:"mammamia" the server will authenticate the client.
The client now can only print by using a token that gets when authenticating.
I removed the fact that the Client class implemented RemoteInterface, I don't think we needed that, the remote object is only the server.
Other things that I forgot...